Conflicts:
	Gemfile.lock
	app/controllers/status_messages_controller.rb
	app/helpers/aspects_helper.rb
	app/views/aspects/index.html.haml
	app/views/layouts/_header.html.haml
	config/routes.rb
	lib/diaspora/user/receiving.rb
	public/javascripts/vendor/fancybox/jquery.fancybox-1.3.1.js
	public/stylesheets/vendor/jquery.fancybox-1.3.1.css

+++ lib/tasks/backup.rake
@@ -1,31 +1,42 @@
 namespace :backup do
-  desc "Backup Mongo"
+  desc "Backup Mysql"
   require File.join(Rails.root, 'config', 'initializers', '_load_app_config.rb')
   require 'cloudfiles'
 
-  task :mongo do
-    Rails.logger.info("event=backup status=start type=mongo")
-
-    if AppConfig[:cloudfiles_username] && AppConfig[:cloudfiles_api_key]
+  task :mysql do
+    NUMBER_OF_DAYS = 3
+    puts("event=backup status=start type=mysql")
+    db = YAML::load(File.open(File.join(File.dirname(__FILE__), '..','..', 'config', 'database.yml')))
+    user = db['production']['user']
+    password = db['production']['password']
+    database = db['production']['database']
+    if AppConfig[:cloudfiles_username] && AppConfig[:cloudfiles_api_key] && !user.blank?
       puts "Logging into Cloud Files"
 
       cf = CloudFiles::Connection.new(:username => AppConfig[:cloudfiles_username], :api_key => AppConfig[:cloudfiles_api_key])
-      mongo_container = cf.container("Mongo Backup")
+      mysql_container = cf.container("MySQL Backup")
 
-      puts "Dumping Mongo"
-      `mongodump -o /tmp/backup/mongo`
+      puts "Dumping Mysql"
+      `mkdir -p /tmp/backup/mysql`
+      `mysqldump --user=#{user} --password=#{password} #{database} >> /tmp/backup/mysql/backup.txt `
 
-      tar_name = "mongo_#{Time.now.to_i}.tar"
-      `tar cfP /tmp/backup/#{tar_name} /tmp/backup/mongo`
+      tar_name = "mysql_#{Time.now.to_i}.tar"
+      `tar cfP /tmp/backup/#{tar_name} /tmp/backup/mysql`
 
-      file = mongo_container.create_object(tar_name)
+      file = mysql_container.create_object(tar_name)
 
       if file.write File.open("/tmp/backup/" + tar_name)
-        Rails.logger.info("event=backup status=success type=mongo")
+        puts("event=backup status=success type=mysql")
         `rm /tmp/backup/#{tar_name}`
-        `rm -rf /tmp/backup/mongo/`
+        `rm -rf /tmp/backup/mysql/`
+
+        files = mysql_container.objects
+        files.sort!.pop(NUMBER_OF_DAYS * 24)
+        files.each do |file|
+          mysql_container.delete_object(file)
+        end
       else
-        Rails.logger.info("event=backup status=failure type=mongo")
+        puts("event=backup status=failure type=mysql")
       end
     else
       puts "Cloudfiles username and api key needed"
@@ -33,7 +44,7 @@ namespace :backup do
   end
 
   task :photos do
-    Rails.logger.info("event=backup status=start type=photos")
+    puts("event=backup status=start type=photos")
 
     if AppConfig[:cloudfiles_username] && AppConfig[:cloudfiles_api_key]
       puts "Logging into Cloud Files"
@@ -42,16 +53,23 @@ namespace :backup do
       photo_container = cf.container("Photo Backup")
 
       tar_name = "photos_#{Time.now.to_i}.tar"
-      `tar cfP /tmp/backup/#{tar_name} /usr/local/app/diaspora/public/uploads/images/`
+      `tar cfP /dev/stdout /usr/local/app/diaspora/public/uploads/images/ | split -d -b 4831838208 - /tmp/backup/#{tar_name}`
 
-      file = photo_container.create_object(tar_name)
+      (0..99).each do |n|
+        padded_str = n.to_s.rjust(2,'0')
+        file = photo_container.create_object(tar_name + padded_str)
+        file_path = "/tmp/backup/" + tar_name + padded_str
 
-      if file.write File.open("/tmp/backup/" + tar_name)
-        Rails.logger.info("event=backup status=success type=photos")
-        `rm /tmp/backup/#{tar_name}`
-      else
-        Rails.logger.info("event=backup status=failure type=photos")
+        if File.exists?(file_path)
+          if file.write File.open(file_path)
+            puts("event=backup status=success type=photos")
+          else
+            puts("event=backup status=failure type=photos")
+          end
+          `rm #{file_path}`
+        end
       end
+
     else
       puts "Cloudfiles username and api key needed"
     end
